requirements.txt
anyio==3.6.2
APScheduler==3.9.1
asgiref==3.5.2
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
drf-api-logger==1.0.8
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
psycopg2-binary==2.9.4
PyJWT==2.6.0
python-telegram-bot==20.0a4
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
webencodings==0.5.1
